requirements.txt
Adafruit-Blinka==5.0.1
adafruit-circuitpython-neopixel==6.0.0
adafruit-circuitpython-pypixelbuf==2.0.0
Adafruit-PlatformDetect==2.13.3
Adafruit-PureIO==1.1.5
libusb1==1.8
ndeflib==0.3.3
nfcpy==1.0.3
pkg-resources==0.0.0
pyDes==2.0.1
pyftdi==0.51.2
pygame==1.9.6
PySDL2==0.9.7
pyserial==3.4
pyusb==1.0.2
rpi-ws281x==4.2.4
RPi.GPIO==0.7.0
sysv-ipc==1.0.1
